+++ buildVars.py
@@ -29,14 +29,14 @@ def _(arg: str) -> str:
 
 # Application information
 app_info = {
-	"app_name": "appTemplate",
-	"app_summary": _("Application template"),
+	"app_name": "nvdaStyleChecker,",
+	"app_summary": _("System for code validations"),
 	"app_description": _(
-		"A template for creating Python desktop applications."
+		"System for checking add-on code and performing validation"
 	),
 	"app_version": "1.1.0",
 	"app_author": "Edilberto Fonseca <edilberto.fonseca@outlook.com>",
-	"spec_file": "install/pyinstaller/appTemplate.spec",
+	"spec_file": "install/pyinstaller/nvdaStyleChecker.spec",
 	"script_name": "src/main.py",
 }